requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
langchain-community
langchain-openai
python-appwrite
requests